backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# TTS
gTTS==2.5.0
pyttsx3==2.90

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0
faker==22.0.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Celery & Queue
celery==5.3.6
redis==5.0.1

# LLM & Encryption
openai==1.10.0
google-generativeai==0.3.2
cryptography==42.0.1

# Audio Processing
openai-whisper==20231117
librosa==0.10.1
praat-parselmouth==0.4.3
